+++ pkg/test/framework/components/istio/version.go
@@ -1,4 +1,4 @@
-package resource
+package istio
 
 import (
 	"strings"
