+++ pkg/test/framework/components/istio/version_test.go
@@ -1,4 +1,4 @@
-package resource
+package istio
 
 import (
 	"fmt"
