+++ pkg/test/framework/resource/settings.go
@@ -19,6 +19,8 @@ import (
 	"path"
 	"strings"
 
+	"istio.io/istio/pkg/test/framework/components/istio"
+
 	"github.com/google/uuid"
 
 	"istio.io/istio/pkg/test/framework/label"
@@ -83,7 +85,7 @@ type Settings struct {
 
 	// Versions are the Istio versions accessible to each of the clusters.
 	// represented as comma-separated versions, such as "1.7.3,1.8.2,1.9.0"
-	Versions Versions
+	Versions istio.Versions
 }
 
 // RunDir is the name of the dir to output, for this particular run.
